+++ CHANGELOG.md
@@ -22,6 +22,12 @@ This file is used to list changes made in each version of the postgresql cookboo
 
 - Remove deprecated tests
 
+## v6.1.2 (2018-04-16)
+
+**this will be the last release of the 6.0 series before all recipes are removed from the cookbook**
+
+- Deprecate all recipes
+
 ## v6.1.1 (2017-03-08)
 
 - Fix pg gem installation on non-omnibus chef runs